Add teleportation back and forth
Now able to touch the ball again in the portal/virtual scene to teleport back!!!
Assign the portal to minimap on grab, bug fixed. Now each crystal ball with teleportation spot works properly.
Hide the in-scene teleport marker when player is in the scene.
Reset the minimap marker on release.